+++ src/rp-download/CacheKeyVerify.cpp
@@ -2,7 +2,7 @@
  * ROM Properties Page shell extension. (rp-download)                      *
  * CacheKeyVerify.cpp: Cache key verifier.                                 *
  *                                                                         *
- * Copyright (c) 2016-2024 by David Korth.                                 *
+ * Copyright (c) 2016-2025 by David Korth.                                 *
  * SPDX-License-Identifier: GPL-2.0-or-later                               *
  ***************************************************************************/
 
@@ -52,6 +52,7 @@ CacheKeyError verifyCacheKey(tstring &outURL, bool &check_newer, const TCHAR *ca
 	// - plus4:  https://rpdb.gerbilsoft.com/plus4/[key]
 	// - ps1:    https://rpdb.gerbilsoft.com/ps1/[key]
 	// - ps2:    https://rpdb.gerbilsoft.com/ps2/[key]
+	// - nes:    https://rpdb.gerbilsoft.com/nes/[key]
 	// - sys:    https://rpdb.gerbilsoft.com/sys/[key] [system info, e.g. update version]
 	const TCHAR *slash_pos = _tcschr(cache_key, _T('/'));
 	if (slash_pos == nullptr || slash_pos == cache_key ||
@@ -146,7 +147,8 @@ CacheKeyError verifyCacheKey(tstring &outURL, bool &check_newer, const TCHAR *ca
 				    !_tmemcmp_inline(cache_key, _T("32x"), 3) ||
 				    !_tmemcmp_inline(cache_key, _T("c64"), 3) ||
 				    !_tmemcmp_inline(cache_key, _T("ps1"), 3) ||
-				    !_tmemcmp_inline(cache_key, _T("ps2"), 3))
+				    !_tmemcmp_inline(cache_key, _T("ps2"), 3) ||
+				    !_tmemcmp_inline(cache_key, _T("nes"), 3))
 				{
 					ok = true;
 				}