+++ NEWS.md
@@ -27,6 +27,10 @@
     handle older Windows disc images that use a long filename for the icon,
     and the disc is authored with Joliet for long filenames but an old version
     of ISO-9660, resulting in 8.3 filenames in the ISO-9660 directories.
+  * NES: Added title screen thumbnailing. Since NES ROMs generally don't have
+    a header with an identifiable game title or serial number, the mapper,
+    CRC32 of the first 8k of PRG ROM, and CRC32 of the first 8k of CHR ROM
+    (if present) are used.
 
 * Bug fixes:
   * Windows: Work around a potential libpng crash when attempting to read